+++ server/src/math/sinogram/utils.py
@@ -8,9 +8,9 @@ def create_offset(array: ndarray[Shape, number], offset: ndarray[Shape, number])
   return (array.T + offset).T
 
 def calculate_emiter_position(radius: float, rotation: float) -> ndarray[(1, 2), number]:
-  return clip(radius * array([cos(rotation), sin(rotation)]), -radius + 1, radius - 1)
+  return clip(radius * array([cos(rotation), sin(rotation)], dtype=float), -radius + 1, radius - 1)
 
 def calculate_detection_positions(radius: float, rotation: float, spread: float, detectors: int) -> \
     ndarray[(int, 2), number]:
   angles = rotation + pi - spread / 2 + arange(detectors) * spread / (detectors - 1)
-  return clip(radius * array([cos(angles), sin(angles)]), -radius + 1, radius - 1)
+  return clip(radius * array([cos(angles), sin(angles)], dtype=float), -radius + 1, radius - 1)
